requirements.txt
requests
python-dotenv
openai
anthropic
google-generativeai
